+++ src/main/scala/uk/gov/hmrc/hmrcfrontend/views/html/components/package.scala
@@ -17,7 +17,7 @@
 package uk.gov.hmrc.hmrcfrontend.views
 package html
 
-import uk.gov.hmrc.govukfrontend.views.html.components.{GovukErrorMessage, GovukHint, GovukLabel, GovukPhaseBanner}
+import uk.gov.hmrc.govukfrontend.views.html.components.{GovukErrorMessage, GovukHint, GovukLabel, GovukPhaseBanner, GovukTextarea}
 
 package object components extends Utils with Aliases {
 
@@ -77,4 +77,9 @@ package object components extends Utils with Aliases {
   type HmrcCurrencyInput = hmrcCurrencyInput
   @deprecated(message = "Use DI", since = "Play 2.6")
   lazy val HmrcCurrencyInput = new hmrcCurrencyInput(GovukErrorMessage, GovukHint, GovukLabel)
+
+  type HmrcCharacterCount = hmrcCharacterCount
+  @deprecated(message = "Use DI", since = "Play 2.6")
+  lazy val HmrcCharacterCount = new hmrcCharacterCount(GovukTextarea, GovukHint)
+
 }